Move SpecimensAction and webPart.jsp to specimen-main
